[SPIRV] Add intrinsics to implement vk::ext_instruction
In HLSL, the `vk::ext_instruction` attribute is used to indicate that a
function represent a particular SPIR-V instruction. The opcode for the
instruction is a parameter to the function.

The input to the backend is a declaration of a function with an
attribute indicating which opcode and extended instruction set to use.
The extended instruciton set can be empty, indicating that a core
instruction is being used.
Then calls to the function are replaced by a single SPIR-V
instruction.

When outputing assembly, the string for the opcode is not readily
available, so we to use `!val` to fill in when the string is not known. This syntax is
accepted by `spirv-as`.

This still does not work when the extended instruction set is unknown to
the backend. This should be improved in a follow up PR.