[SPIRV] Add intrinsics to implement vk::ext_instruction
In HLSL, the `vk::ext_instruction` attribute is used to indicate that a
function represent a particular SPIR-V instruction. The opcode for the
instruction is a parameter to the function.

To implement the, we add a pseudo instuction that will be emitted in the
spirv using a special case in the code generator.

When outputing assembly, the string for the opcode is not readily
available, so we oped to output a standin, `OpUnknown(val)`. This will
not assemble, but will allow the user to see their code.

If we wanted to improve this, we would have a lookup table for known
instruction, but we will always have to have the fall back. This feature
is intended to be used to implement new SPIR-V features like vendor
extensions without having to update the compiler.

# Conflicts:
#	llvm/lib/Target/SPIRV/SPIRVInstructionSelector.cpp